Executable
Successfully generated a working executable with all of my current work. Had to fix a couple of things. 1. AWT isn't compatible with gluon so had to open links with an even easier method that wasn't shown to me the other day... 2. The items.xml file is now outside of the exe and I plan to bundle them in a zip folder for releases.