Item display finished
Always reluctant to use the f word, but for now item display is looking really good. Properly loads and unloads cards for performance and have fixed the few bugs I found with it. I really need to move on to finishing up the banner and getting all XML info finished up.