Item loading and saving rework
Reworked the way items are loaded in and how obtained/not obtained status is saved. Now an infinite number of XML files can be used to load in items by naming them something that includes "_items" and ".xml". Currently save.xml is used to indicate if the item is obtained or not. save.xml will contain a node for the item it it's been ovtained.