Images, Lootlemon, and BL4
Implemented buttons for Lootlemon links and finished up the rest of the custom images. Also implemented a button to reset the save state, requiring 3 presses of the button to confirm. Also got all current info for BL4 items into the xml. Need to do some testing but this may be good for a 1.0 release.